--- modules/book.js
@@ -1,7 +1,0 @@
-export default class Book {
-  constructor(title, author, id) {
-    this.title = title;
-    this.author = author;
-    this.id = id;
-  }
-}
